Remove sidebar icon size setting & allow iCloud
The sidebar icon size also changes the icon size in the system settings. I do not mind the size change in finder however the font size becomes a lot smaller in the system settings side bar and that makes it a lot harder to read.

I feel that removing iCloud by default is a little harsh and should be moved a separate script. Apple is all about ecosystem so while some people might not use iCloud disabling this ecosystem feature by default is a very opinionated move.